Break down huge file
