+++ packages/alfa-media/src/condition/index.ts
@@ -1,0 +1,1 @@
+export * from "./condition";
