+++ packages/alfa-media/tsconfig.json
@@ -2,12 +2,15 @@
   "$schema": "http://json.schemastore.org/tsconfig",
   "extends": "../tsconfig.json",
   "files": [
+    "src/condition/condition.ts",
+    "src/condition/index.ts",
     "src/feature/feature.ts",
     "src/feature/index.ts",
     "src/index.ts",
+    "src/list.ts",
     "src/matchable.ts",
-    "src/media.ts",
     "src/modifier.ts",
+    "src/query.ts",
     "src/resolver.ts",
     "src/type.ts",
     "src/value/index.ts",
