+++ packages/alfa-rules/src/sia-r83/rule.ts
@@ -14,10 +14,7 @@ import {
 } from "@siteimprove/alfa-dom";
 import { Hash } from "@siteimprove/alfa-hash";
 import { Iterable } from "@siteimprove/alfa-iterable";
-import {
-  Feature as MediaFeature,
-  Value as MediaValue,
-} from "@siteimprove/alfa-media";
+import { Media } from "@siteimprove/alfa-media";
 import { None, Option } from "@siteimprove/alfa-option";
 import { Predicate } from "@siteimprove/alfa-predicate";
 import { Refinement } from "@siteimprove/alfa-refinement";
@@ -32,8 +29,8 @@ import { expectation } from "../common/act/expectation";
 
 import { Scope, Stability } from "../tags";
 
-const { isHeight, isWidth } = MediaFeature;
-const { Discrete, Range } = MediaValue;
+const { isHeight, isWidth } = Media.Feature;
+const { Discrete, Range } = Media.Value;
 
 const { or, not, equals } = Predicate;
 const { and, test } = Refinement;
@@ -527,8 +524,8 @@ function usesMediaRule(
  * We currently do not support calculated media queries. But this is lost in the
  * typing of Media.Feature. Here, we simply consider them as "good" (font relative).
  */
-function isFontRelativeMediaRule<F extends MediaFeature>(
-  refinement: Refinement<MediaFeature, F>,
+function isFontRelativeMediaRule<F extends Media.Feature>(
+  refinement: Refinement<Media.Feature, F>,
 ): Predicate<MediaRule> {
   return (rule) =>
     Iterable.some(rule.queries.queries, (query) =>
@@ -554,9 +551,9 @@ function isFontRelativeMediaRule<F extends MediaFeature>(
     );
 }
 
-function usesFontRelativeMediaRule<F extends MediaFeature>(
+function usesFontRelativeMediaRule<F extends Media.Feature>(
   device: Device,
-  refinement: Refinement<MediaFeature, F>,
+  refinement: Refinement<Media.Feature, F>,
   context: Context = Context.empty(),
 ): Predicate<Element> {
   return usesMediaRule(isFontRelativeMediaRule(refinement), device, context);
